pyproject.toml
[project]
name = "blog-translator"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "pislashpi", email = "123241027+pislashpi@users.noreply.github.com" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "feedparser>=6.0.11",
    "requests>=2.32.3",
    "google-generativeai>=0.8.4",
    "google-genai>=1.3.0",
    "beautifulsoup4>=4.13.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/blog_translator"]
